Update README.md
Systems Engineer passionate about information security and process automation. Specialized in server administration in enterprise environments, with a proactive approach towards continuous improvement and operational efficiency. Committed to the implementation of robust solutions that guarantee the integrity and confidentiality of data. Constantly exploring new technologies to stay up to date in a constantly evolving digital world.